# public_api_swagger_openapi_2.0.yaml
swagger: '2.0'
info:
  version: 'latest'
  title: Neobookings Public API
  description: 'Neobookings REST API'
host: ws-test.neobookings.com
basePath: "/api/v2/"
schemes:
- https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
- application/json
- application/xml
produces:
- application/json
- application/xml
...
paths:
  "/AuthenticatorRQ":
    post:
      tags:
      - neoPublic
      operationId: AuthenticatorRQ
      produces:
      - application/json
      - application/xml
      parameters:
      - name: Body
        in: body
        required: true
        schema:
          "$ref": "#/definitions/AuthenticatorRQ"
      responses:
        '200':
          description: 'Esquema de respuesta'
          schema:
            "$ref": "#/definitions/AuthenticatorRS"
      x-unitTests:
      - request:
        method: POST
        uri: "/AuthenticatorRQ"
        headers:
          Content-Type: application/json
...
definitions:
  AuthenticatorRQ:
    title: 'AuthenticatorRQ'
    description: 'Mensaje utilizado para autenticarse en el sistema y crear una sesión.'
    type: object
    required: ['Request','Credentials']
    properties:
      Request:
        $ref: '#/definitions/Request'
      Credentials:
        $ref: '#/definitions/Credentials'
  Request:
    title: 'Request'
    description: 'Elemento requerido con los datos obligatorios para todas las solicitudes.
            '
    type: object
    required: ['RequestId','Timestamp','Language']
    properties:
      RequestId:
        description: 'Identificador de solicitud.'
        type: string
      Timestamp:
        description: 'Instante temporal de la solicitud.'
        type: string
      Language:
        description: ''
        type: string
        enum:
        - "aa"
        - "ab"
        - "ae"
        - "af"
        - "ak"
        - "am"
        - "an"
        - "ar"
        - "as"
        - "av"
        - "ay"
        - "az"
        - "ba"
        - "be"
        - "bg"
        - "bh"
        - "bi"
        - "bm"
        - "bn"
        - "bo"
        - "br"
        - "bs"
        - "ca"
        - "ce"
        - "ch"
        - "co"
        - "cr"
        - "cs"
        - "cu"
        - "cv"
        - "cy"
        - "da"
        - "de"
        - "dv"
        - "dz"
        - "ee"
        - "el"
        - "en"
        - "eo"
        - "es"
        - "et"
        - "eu"
        - "fa"
        - "ff"
        - "fi"
        - "fj"
        - "fo"
        - "fr"
        - "fy"
        - "ga"
        - "gd"
        - "gl"
        - "gn"
        - "gu"
        - "gv"
        - "ha"
        - "he"
        - "hi"
        - "ho"
        - "hr"
        - "ht"
        - "hu"
        - "hy"
        - "hz"
        - "ia"
        - "id"
        - "ie"
        - "ig"
        - "ii"
        - "ik"
        - "io"
        - "is"
        - "it"
        - "iu"
        - "ja"
        - "jv"
        - "ka"
        - "kg"
        - "ki"
        - "kj"
        - "kk"
        - "kl"
        - "km"
        - "kn"
        - "ko"
        - "kr"
        - "ks"
        - "ku"
        - "kv"
        - "kw"
        - "ky"
        - "la"
        - "lb"
        - "lg"
        - "li"
        - "ln"
        - "lo"
        - "lt"
        - "lu"
        - "lv"
        - "mg"
        - "mh"
        - "mi"
        - "mk"
        - "ml"
        - "mn"
        - "mr"
        - "ms"
        - "mt"
        - "my"
        - "na"
        - "nb"
        - "nd"
        - "ne"
        - "ng"
        - "nl"
        - "nn"
        - "no"
        - "nr"
        - "nv"
        - "ny"
        - "oc"
        - "oj"
        - "om"
        - "or"
        - "os"
        - "pa"
        - "pi"
        - "pl"
        - "ps"
        - "pt"
        - "qu"
        - "rm"
        - "rn"
        - "ro"
        - "ru"
        - "rw"
        - "sa"
        - "sc"
        - "sd"
        - "se"
        - "sg"
        - "si"
        - "sk"
        - "sl"
        - "sm"
        - "sn"
        - "so"
        - "sq"
        - "sr"
        - "ss"
        - "st"
        - "su"
        - "sv"
        - "sw"
        - "ta"
        - "te"
        - "tg"
        - "th"
        - "ti"
        - "tk"
        - "tl"
        - "tn"
        - "to"
        - "tr"
        - "ts"
        - "tt"
        - "tw"
        - "ty"
        - "ug"
        - "uk"
        - "ur"
        - "uz"
        - "ve"
        - "vi"
        - "vo"
        - "wa"
        - "wo"
        - "xh"
        - "yi"
        - "yo"
        - "za"
        - "zh"
        - "zu"
  Credentials:
    title: 'Credentials'
    description: 'Credenciales del usuario que solicita autentificarse.'
    type: object
    required: ['ClientCode','SystemCode','Username','Password']
    properties:
      ClientCode:
        description: 'Código del cliente.'
        type: string
      SystemCode:
        description: 'Código del sistema.'
        type: string
      Username:
        description: 'Nombre de usuario.'
        type: string
      Password:
        description: 'Contraseña de usuario.'
        type: string
  AuthenticatorRS:
    title: 'AuthenticatorRS'
    description: 'Respuesta de autenticación en el sistema con el identificador de sesión.'
    type: object
    required: ['Response']
    properties:
      Response:
        $ref: '#/definitions/Response'
      Token:
        description: 'Identificador de sesión'
        type: string
  Response:
    title: 'Response'
    description: 'Elemento requerido con los datos de la respuesta.'
    type: object
    required: ['StatusCode','RequestId','Timestamp','TimeResponse']
    properties:
      StatusCode:
        description: 'Codigo de estado.'
        type: number
      RequestId:
        description: 'Identificador de solicitud.'
        type: string
      Timestamp:
        description: 'Instante temporal de la respuesta.'
        type: string
      TimeResponse:
        description: 'Tiempo de respuesta.'
        type: number
      Success:
        $ref: '#/definitions/Success'
      Warning:
        type: array
        items:
          $ref: '#/definitions/Warning'
      Error:
        type: array
        items:
          $ref: '#/definitions/Error'
  Success:
    title: 'Success'
    description: 'Mensaje de éxito.'
    type: object
  Warning:
    title: 'Warning'
    description: 'Mensaje de advertencia.'
    type: object
    required: ['Code','Description']
    properties:
      Code:
        description: 'Código de la advertencia.'
        type: string
      Description:
        description: 'Descripción de la advertencia.'
        type: string
  Error:
    title: 'Error'
    description: 'Mensaje de error.'
    type: object
    required: ['Code','Description']
    properties:
      Code:
        description: 'Código del error.'
        type: string
      Description:
        description: 'Descripción del error.'
        type: string
...